Flytter en del av logikken fra EmployeeHourlySalaryStorage og OvertimePayoutStorage ut i SalaryService
